// zenstack/schema.ts
//////////////////////////////////////////////////////////////////////////////////////////////
// DO NOT MODIFY THIS FILE                                                                  //
// This file is automatically generated by ZenStack CLI and should not be manually updated. //
//////////////////////////////////////////////////////////////////////////////////////////////

/* eslint-disable */

import { type SchemaDef, type AttributeApplication, type FieldDefault, ExpressionUtils } from "@zenstackhq/schema";
export class SchemaType implements SchemaDef {
    provider = {
        type: "postgresql"
    } as const;
    models = {
        ContactSubmission: {
            name: "ContactSubmission",
            fields: {
                id: {
                    name: "id",
                    type: "Int",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("autoincrement") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("autoincrement") as FieldDefault
                },
                name: {
                    name: "name",
                    type: "String"
                },
                email: {
                    name: "email",
                    type: "String",
                    attributes: [{ name: "@email" }] as readonly AttributeApplication[]
                },
                reason: {
                    name: "reason",
                    type: "String"
                },
                message: {
                    name: "message",
                    type: "String"
                },
                consent: {
                    name: "consent",
                    type: "Boolean"
                },
                createdAt: {
                    name: "createdAt",
                    type: "DateTime",
                    attributes: [{ name: "@default", args: [{ name: "value", value: ExpressionUtils.call("now") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("now") as FieldDefault
                }
            },
            idFields: ["id"],
            uniqueFields: {
                id: { type: "Int" }
            }
        }
    } as const;
    plugins = {};
}
export const schema = new SchemaType();